Public API

MaximumLikelihoodProblems.fit

MaximumLikelihoodProblems.loglikelihood

MaximumLikelihoodProblems.fit — Function

fit(transformed_gradient_problem, theta_hat_initial; kwargs...)

Find the maximum likelihood estimatator for the parameters theta.

Arguments

transformed_gradient_problem

theta_hat_initial

Keyword Arguments

learning_rate. Default value: 0.0001

max_iterations. Default value: 1000000

show_progress_meter. Default value: true

throw_convergence_exception. Default value: true

tolerance. Default value: 1.0e-10

See the documentation for fully worked-out examples.

MaximumLikelihoodProblems.loglikelihood — Function

loglikelihood(transformed_gradient_problem, theta)

Return the value of the log likelihood function evaluated at theta.

Arguments

transformed_gradient_problem

theta